fix cpplint errors

+++ include/WebApi_ws_vedirect_live.h
@@ -21,6 +21,6 @@ private:
 
     uint32_t _lastWsPublish = 0;
     uint32_t _lastVedirectUpdateCheck = 0;
-    unsigned long _lastWsCleanup = 0;
+    uint32_t _lastWsCleanup = 0;
     uint32_t _newestVedirectTimestamp = 0;
 };
